Rename async_submission.py to submission.py
Rename the async_submisison.py model module to submission.py in order to
make it consistent with the task module name and because it's shorter.
